Rename src/phantom.js -> src/jsdom.js
